# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LLM Proxy HTTP API
  version: 1.0.0
  description: >-
    The sole canonical wire contract for llm-proxy-owned proxy, browser
    configuration, and management endpoints. Each configured text model owns
    one explicit upstream wire contract and execution lifecycle. OpenAI
    Gemini 3.5 Transcribe accepts file dictation through synchronous Interactions.
    The filename extension selects a supported audio MIME type. An unsupported
    extension returns HTTP 400 with invalid_audio_input. The upload byte limit
    applies. Gemini enforces its one-hour audio duration limit.
    Gemini 3.6 Flash accepts reasoning efforts minimal, low, medium, and high.
    Gemini 3.7 Flash accepts low, medium, and high. An omitted effort leaves
    the provider default in effect. Both routes accept text, image, and audio
    input with a maximum of 65,536 output tokens.
    Responses and Gemini 3.x Interactions use stored pollable resources.
    OpenAI-compatible Chat Completions and Anthropic Messages complete
    synchronously. Public proxy operations remain blocking and never expose an
    upstream resource id. OpenAI resource ids remain in memory only, but the
    current adapter does not cancel or delete stored Responses, so OpenAI
    account retention policy applies. Background Gemini interaction ids are
    also request-local; the adapter cancels active resources and deletes every
    stored interaction on exit. Each Gemini Interactions route rejects
...
servers:
...
paths:
...
  /api/management/connections/{connection_id}:
    get:
      operationId: getManagementConnection
      summary: Read a connection with masked credentials
      tags:
        - Management
      security:
        - TAuthSession: []
      parameters:
        - name: connection_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - $ref: '#/components/parameters/OptionalOrigin'
      responses:
        '200':
          description: Operation completed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountConnection'
        '400':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '401':
          $ref: '#/components/responses/ManagementUnauthorized'
        '403':
          $ref: '#/components/responses/ManagementForbidden'
        '404':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '409':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '500':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
    put:
      operationId: replaceManagementConnection
      summary: Update a connection at its expected version
      tags:
        - Management
      security:
        - TAuthSession: []
      x-llm-proxy-required-request-content-type: application/json
      parameters:
        - name: connection_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - $ref: '#/components/parameters/OptionalOrigin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountConnectionWrite'
      responses:
        '200':
          description: Operation completed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountConnection'
        '400':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '401':
          $ref: '#/components/responses/ManagementUnauthorized'
        '403':
          $ref: '#/components/responses/ManagementForbidden'
        '404':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '409':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '415':
          $ref: '#/components/responses/ManagementUnsupportedMediaType'
        '422':
          $ref: '#/components/responses/ProviderKeyRejected'
        '429':
          $ref: '#/components/responses/ProviderKeyVerificationRateLimited'
        '500':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '503':
          $ref: '#/components/responses/ProviderKeyVerificationUnavailable'
        '504':
          $ref: '#/components/responses/ProviderKeyVerificationTimedOut'
    delete:
      operationId: deleteManagementConnection
      summary: Delete an unassigned connection
      tags:
        - Management
      security:
        - TAuthSession: []
      x-llm-proxy-required-request-content-type: application/json
      parameters:
        - name: connection_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
        - $ref: '#/components/parameters/OptionalOrigin'
      responses:
        '204':
          description: Operation completed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '401':
          $ref: '#/components/responses/ManagementUnauthorized'
        '403':
          $ref: '#/components/responses/ManagementForbidden'
        '404':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '409':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '415':
          $ref: '#/components/responses/ManagementUnsupportedMediaType'
        '500':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
...
  /api/management/tenants/{tenant_id}/connections/{provider}:
    put:
      operationId: assignManagementConnection
      summary: Select an account connection or hosted access grant for one provider
      description: >-
        The resource must belong to the tenant's account. A hosted grant must also name this tenant and provider.
        Detach the current assignment before selecting a different resource. The service does not substitute credential sources.
      tags:
        - Management
      security:
        - TAuthSession: []
      x-llm-proxy-required-request-content-type: application/json
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ProviderPath'
        - $ref: '#/components/parameters/OptionalOrigin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantConnectionAssignment'
      responses:
        '200':
          description: Operation completed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantProfile'
        '400':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '401':
          $ref: '#/components/responses/ManagementUnauthorized'
        '403':
          $ref: '#/components/responses/ManagementForbidden'
        '404':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '409':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '415':
          $ref: '#/components/responses/ManagementUnsupportedMediaType'
        '500':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
    delete:
      operationId: detachManagementConnection
      summary: Detach the selected provider resource and optionally clear dependent defaults
      tags:
        - Management
      security:
        - TAuthSession: []
      x-llm-proxy-required-request-content-type: application/json
      parameters:
        - $ref: '#/components/parameters/TenantID'
        - $ref: '#/components/parameters/ProviderPath'
        - name: clear_defaults
          in: query
          schema:
            type: string
            enum:
              - 'true'
              - 'false'
          description: Use true to confirm removal of dependent default routes.
        - $ref: '#/components/parameters/OptionalOrigin'
      responses:
        '204':
          description: Operation completed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
        '400':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '401':
          $ref: '#/components/responses/ManagementUnauthorized'
        '403':
          $ref: '#/components/responses/ManagementForbidden'
        '404':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '409':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
        '415':
          $ref: '#/components/responses/ManagementUnsupportedMediaType'
        '500':
          description: The connection operation failed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionError'
...
components:
  securitySchemes:
    MCPAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: TAuth OAuth access token with the API audience and llm-proxy:use scope.
    TenantBearerKey:
      type: http
      scheme: bearer
      bearerFormat: tenant-client-key
    TenantClientKey:
      type: apiKey
      in: query
      name: key
      description: The tenant client secret generated by the management API.
    TAuthSession:
      type: apiKey
      in: cookie
      name: app_session_llm_proxy
      description: The exact deployment-configured TAuth session cookie. The production cookie name is configured through management.session_cookie_name.
  parameters:
...
    TenantID:
      name: tenant_id
      in: path
      required: true
      description: Opaque tenant identifier from the authenticated account response.
      schema:
        type: string
        minLength: 1
    ProviderPath:
      name: provider
      in: path
      required: true
      description: Canonical provider identifier from the public capability catalog, such as zai. Retired zhipu and glm values are invalid.
      schema:
        type: string
        minLength: 1
...
    OptionalOrigin:
      name: Origin
      in: header
      required: false
      description: When supplied, must exactly match the configured public origin for mutations. Matching origins receive credentialed CORS headers.
      schema:
        type: string
        format: uri
...
  responses:
...
    ManagementUnauthorized:
      description: The TAuth session is absent or invalid.
    ManagementForbidden:
      description: The user lacks the required role or the supplied Origin is not allowed.
...
    ManagementUnsupportedMediaType:
      description: A management mutation did not supply application/json.
    ProviderKeyRejected:
      description: The candidate credential was rejected or cannot access the selected text model. Nothing was saved.
      content:
        text/plain:
          schema:
            type: string
            const: provider_key_rejected
    ProviderKeyVerificationRateLimited:
      description: The provider rate-limited the single verification operation. The credential remains unconfirmed and nothing was saved.
      content:
        text/plain:
          schema:
            type: string
            const: provider_key_verification_rate_limited
    ProviderKeyVerificationUnavailable:
      description: The provider was unavailable or returned a malformed success envelope. The credential remains unconfirmed and nothing was saved.
      content:
        text/plain:
          schema:
            type: string
            const: provider_key_verification_unavailable
    ProviderKeyVerificationTimedOut:
      description: The single verification operation timed out or was canceled. The credential remains unconfirmed and nothing was saved.
      content:
        text/plain:
          schema:
            type: string
            const: provider_key_verification_timed_out
...
  headers:
...
    NoStore:
      description: Prevents persistence of authenticated or secret-bearing responses.
      schema:
        type: string
        const: no-store
      x-required: true
...
  schemas:
    AccountConnectionWrite:
      type: object
      additionalProperties: false
      required:
        - name
        - provider
        - fields
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        provider:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
        version:
          type: integer
          minimum: 0
          description: Use zero or omit on creation. Updates require the current connection version.
    AccountConnection:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - provider
        - version
        - fields
        - tenant_ids
        - created_at
        - updated_at
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        version:
          type: integer
          minimum: 1
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ProviderField'
        tenant_ids:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TenantConnectionAssignment:
      type: object
      additionalProperties: false
      required:
        - kind
        - resource_id
      properties:
        kind:
          $ref: '#/components/schemas/ProviderAssignmentKind'
        resource_id:
          type: string
          minLength: 1
    ProviderAssignmentKind:
      type: string
      enum: [account_connection, hosted_access_grant]
...
    ConnectionError:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
          properties:
            code:
              type: string
...
    TenantProfile:
      type: object
      additionalProperties: false
      required:
        - tenant
        - providers
        - proxy
      properties:
        tenant:
          $ref: '#/components/schemas/Tenant'
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderProfile'
        proxy:
          $ref: '#/components/schemas/ProxyPaths'
...